+++ Transfer/ViewControllers/IntroViewController.m
@@ -232,9 +232,9 @@
     }
 
     [self setReportedPage:currentPage];
-    MCLog(@"Report page:%ld", currentPage);
+    MCLog(@"Report page:%ld", (long)currentPage);
 
-    [[GoogleAnalytics sharedInstance] sendAppEvent:@"IntroScreensSlided" withLabel:[NSString stringWithFormat:@"%ld", currentPage + 1]];
+    [[GoogleAnalytics sharedInstance] sendAppEvent:@"IntroScreensSlided" withLabel:[NSString stringWithFormat:@"%d", currentPage + 1]];
 }
 
 - (NSAttributedString *)attributedMessage:(NSString *)message bold:(NSArray *)boldTexts {
